Subjective Community Well-Being Dataset (25 Variables)

These Subjective Community Well-Being (SCWB) data are a random subset from the full dataset used in the paper by Padgett et al. (in press). The SCWB dataset contains 501 responses to the 20 SCWB items plus 5 pre-computed total scores on the five theorized domains of Good Relationships, Proficient Leadership, Healthy Practices, Satisfying Community, and Strong Mission.

Usage

data(SCWB)

Format

SCWB A data frame with 501 observations of 25 variables.

cwbcloserel

cwbrespect

cwbtrust

cwbmutuality

cwbbeneficence

cwbintegrity

cwbcompetence

cwbvision

cwbrelgrowth

cwbfairness

cwbsustenance

cwbachieve

cwbsatisf

cwbvalue

cwbbelonging

cwbwelcome

cwbpurpose

cwbcontrib

cwbinterconn

cwbsynergy

total_goodrel

Mean of items cwbcloserel, cwbrespect, cwbtrust, cwbmutuality

total_prolead

Mean of items cwbbeneficence, cwbintegrity, cwbcompetence, cwbvision

total_healthyprac

Mean of items cwbrelgrowth, cwbfairness, cwbsustenance, cwbachieve

total_satcomm

Mean of items cwbsatisf, cwbvalue, cwbbelonging, cwbwelcome

total_strongmiss

Mean of items cwbpurpose, cwbcontrib, cwbintercon, cwbsynergy

Source

Data used in Padgett et al. (in press).

References

Padgett, R.N., ...

Examples

# SCWB example used in preprint
data(SCWB)