REC Metric 2

For each domain i, the average of the REC entries within domain vs. the average of the REC entries when looking at correlations between indicators in domain D=i with indicators in domains other than i (note that this does NOT include correlations in domain j with other indicators in domain j; for a given domain, it is basically the average of the REC entries within the domain vs. the average of the REC entries for all of the other entries in the column of that domain). We might call this the REC domain coherence for domain i

Usage

rec.avg.wb.domains(data, domainlist = NULL, domainnames = NULL, na.rm = TRUE)

Arguments

data

A data.frame or tibble (required)

domainlist

A list of character vectors or numeric vectors denoting the columns for each domain (required)

domainnames

A character vector supplying updated names to be attached to output

na.rm

logical (defaults to TRUE)

Value

A single numerical value

Details

Examples

  # Use the SCWB data example
  mylist <- list(
    domain_goodrel = c("cwbcloserel", "cwbrespect", "cwbtrust", "cwbmutuality"),
    domain_prolead = c("cwbbeneficence", "cwbintegrity", "cwbcompetence", "cwbvision"),
    domain_healthyprac = c("cwbrelgrowth", "cwbfairness", "cwbsustenance", "cwbachieve"),
    domain_satcomm = c("cwbsatisf", "cwbvalue", "cwbbelonging", "cwbwelcome"),
    domain_strongmiss = c("cwbpurpose", "cwbcontrib", "cwbinterconn", "cwbsynergy")
  )
  rec.avg.wb.domains(SCWB, mylist)
#>     Domain Avg_Rec_Within_Domain Avg_REC_Between_Domains
#> 1 Domain 1            0.19081023             -0.03573701
#> 2 Domain 2            0.17721910             -0.03152309
#> 3 Domain 3            0.10720594             -0.01796224
#> 4 Domain 4            0.08741102             -0.01522356
#> 5 Domain 5            0.14737622             -0.03268333